Include the initial sequence in the trajectory.
Previously I was assuming that the first "current_seq" in the trajectory
was the initial sequence, but I realized that this is only true if the
first move is rejected, which it usually isn't.  Now the correct initial
sequence is included in the trajectory itself.

The *.tsv file format is getting a little hacky.  It would be better to
use the *.hdf format, which could've handled this extra metadata without
any contortions.  But switching wouldn't be worth the time now.