Add the "active macrostate" score term.
This was a fairly major change, because I had to add an "active
constraints" member to the construct class and I had to generalize
ViennaRnaFold to accommodate the constrained calculations.

The test cover the most crucial parts of the code, but they aren't all
the way there yet.  The score term itself is not tested, although it
empirically seems to work well.  The sequence-modifying methods of the
Domain class aren't tested and don't work, because they don't update the
active-state constraint.